+++ packages/shared-base/src/index.ts
@@ -2,29 +2,21 @@ export * from '@masknet/base'
 
 export * from './constants.js'
 export * from './types.js'
+export * from './types/index.js'
 export * from './helpers/index.js'
 
-export * from './Account/index.js'
-export * from './MIME/index.js'
 export * from './Messages/Events.js'
 export * from './Messages/CrossIsolationEvents.js'
 export * from './Messages/index.js'
-export * from './NameService/index.js'
 export * from './NextID/index.js'
 export * from './NextID/types.js'
 export * from './Pageable/index.js'
-export * from './Persona/types.js'
-export * from './Plugin/index.js'
-export * from './Post/type.js'
 export * from './Routes/index.js'
 export * from './Site/index.js'
 export * from './Site/types.js'
 export * from './Sniffings/index.js'
-export * from './Social/index.js'
 export * from './Subscriptions/index.js'
-export * from './Wallet/index.js'
 export * from './collections/index.js'
-export * from './color/index.js'
 export * from './i18n/index.js'
 export * from './kv-storage/index.js'
 export * from './onDemandWorker/index.js'
