+++ packages/shared-base/src/types/index.ts
@@ -1,0 +1,9 @@
+export * from './Account.js'
+export * from './Color.js'
+export * from './PluginID.js'
+export * from './Persona.js'
+export * from './Post.js'
+export * from './MIME.js'
+export * from './NameService.js'
+export * from './Wallet.js'
+export * from './SocialIdentity.js'
